Split the invoicing service in a smaller module
